requirements.txt
matplotlib==3.7.1
numpy==1.25.0
pandas==2.0.2
Pillow==9.5.0
Pillow==9.5.0
plotly==5.15.0
seaborn==0.12.2
streamlit==1.23.1
